requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.4
gunicorn==23.0.0
cloudinary==1.41.0

# === Tambahan untuk Google Drive API ===
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1